Adding Explicit Opperator To The Vector3 And Fixing Vector2

+++ Struct/Vector2.cs
@@ -8,7 +8,7 @@ namespace MeshToolbox
 {
     public struct Vector2
     {
-        public Vector2(float X, float Y)
+        public Vector2(double X, double Y)
         {
             x = X;
             y = Y;
@@ -17,7 +17,7 @@ namespace MeshToolbox
         public static Vector2 zero = new Vector2(0, 0);
         public static Vector2 one = new Vector2(1, 1);
 
-        public float x;
-        public float y;
+        public double x;
+        public double y;
     }
 }
